Move models into more easily importable files
